+++ CHANGELOG.md
@@ -10,9 +10,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * Update parser.
   * Strings without quote will be treated as string now.
   * Variables can be parsed now. The default value will be used.
+* Focus tree preview will show warnings for invalid code.
 
 ### Fixed
 * An issue that preview button shows incorrect type of preview.
+* An issue that in some case focus tree preview shows nothing.
 
 ## [0.3.4] - 2020/06/23 - Latest